feat(api): add family-neutral interface address configs
`FlatInterfaceConfig` still sends its IPv4 address data through scalar fields, with IPv6 added later as a separate sidecar. That shape cannot represent an IPv6-only interface and complicates the planned family-neutral migration.

Add `InterfaceAddressConfig` and an additive `addresses` field, then have Core mirror the existing admin and tenant values into it in stable IPv4-then-IPv6 order. The legacy scalars and IPv6 sidecar stay populated, so older agents keep working while #2395 moves readers over.

This intentionally does not change agent rendering or remove the IPv4 requirement; those remain #2395 and #2402.

Tests added!

This supports https://github.com/NVIDIA/infra-controller/issues/2394
