+++ crates/rpc/proto/forge.proto
@@ -4471,16 +4471,13 @@ message DhcpDiscovery {
   optional bytes duid = 10;
 }
 
-// DHCP protocol family used to validate DhcpDiscovery.message_kind and duid.
-// Field absence means legacy inference; *_UNSPECIFIED means an explicit invalid
-// value and is rejected when supplied.
+// IP address family.
 enum AddressFamily {
-  // Invalid when explicitly supplied. Omit DhcpDiscovery.address_family instead
-  // when using legacy family inference.
+  // No address family selected. Invalid in fields that require an explicit family.
   ADDRESS_FAMILY_UNSPECIFIED = 0;
-  // DHCPv4. Compatible only with MESSAGE_KIND_V4_DISCOVER and no DUID.
+  // IPv4.
   ADDRESS_FAMILY_V4 = 1;
-  // DHCPv6. Compatible only with V6 MessageKind values and requires a DUID.
+  // IPv6.
   ADDRESS_FAMILY_V6 = 2;
 }
 
@@ -5009,6 +5006,12 @@ message FlatInterfaceConfig {
   // Interface-local routing profile details. These narrow the VPC routing
   // profile for this interface.
   optional FlatInterfaceRoutingProfile interface_routing_profile = 21;
+
+  // Family-neutral replacement for gateway, ip, interface_prefix, prefix, svi_ip,
+  // and ipv6_interface_config. An empty list identifies a legacy payload. Writers
+  // must emit at most one entry per family, ordered V4 before V6. Readers must
+  // select entries by address_family rather than position.
+  repeated InterfaceAddressConfig addresses = 22;
 
   // The details of the network security group associated with
   // either the instance or its parent VPC.
@@ -10180,4 +10183,23 @@ message SitePrefixIdList {
 message SitePrefixList {
   option (carbide.codegen.v1.message_derive) = "serde::Serialize";
   repeated SitePrefix site_prefixes = 1;
+}
+
+// Family-specific address configuration for a managed-host interface.
+message InterfaceAddressConfig {
+  option (carbide.codegen.v1.message_derive) = "serde::Serialize";
+
+  // Family shared by every address and prefix in this entry. Must be V4 or V6.
+  AddressFamily address_family = 1;
+  // IPv4 gateway CIDR. For IPv6, the DPU-side port/linknet CIDR used as
+  // gateway_cidr, matching FlatInterfaceIpv6Config.interface_prefix.
+  string gateway = 2;
+  // Host address, without a prefix length.
+  string ip = 3;
+  // Prefix allocated to this interface, such as an FNN linknet or host route.
+  string interface_prefix = 4;
+  // Prefix for the network segment that contains this interface.
+  string prefix = 5;
+  // DPU-hosted SVI address used as the host gateway when configured.
+  optional string svi_ip = 6;
 }
